requirements/base.txt
asgiref
black
click
coverage
django
django-filter
exceptiongroup
factory-boy
faker
flake8
inflection
iniconfig
mccabe
mypy-extensions
packaging
pathspec
platformdirs
pluggy
pycodestyle
pyflakes
pytest
pytest-cov
pytest-django
pytest-factoryboy
pytest-mock
python-dateutil
python-environ
six
sqlparse
tomli
typing-extensions
whitenoise